sepereated screens from the main page file. nice and clean now
